Moved tile entities to block/entities package
