+++ src/main/java/com/drakmyth/minecraft/manufactory/blocks/entities/GrinderTileEntity.java
@@ -3,7 +3,7 @@
  *  Copyright (c) 2020 Drakmyth. All rights reserved.
  */
 
-package com.drakmyth.minecraft.manufactory.tileentities;
+package com.drakmyth.minecraft.manufactory.blocks.entities;
 
 import java.util.ArrayList;
 import java.util.List;
